+++ src/extension.ts
@@ -108,6 +108,7 @@ function registerCommands(
     ['clawAI.sendFeedback', () => coordinator.commands.sendFeedback()],
     ['clawAI.searchRunHistory', () => coordinator.commands.searchRunHistory()],
     ['clawAI.showUsage', () => coordinator.commands.showUsage()],
+    ['clawAI.askSideQuestion', () => coordinator.commands.askSideQuestion()],
     ['clawAI.compactConversation', () => coordinator.commands.compactConversation()],
     ['clawAI.selectOutputStyle', () => coordinator.commands.selectOutputStyle()],
     ['clawAI.toggleFocusView', () => coordinator.commands.toggleFocusView()],